feature: move nutriments arraus out of the component. Refresh the table when click SKIP or VALIDATE
