+++ src/pages/nutrition/nutritionFields.js
@@ -1,0 +1,55 @@
+const basicNutriments = [
+  {
+    off_nutriment_id: "energy_kj",
+    label: "Energie (kJ)",
+    value: "",
+    unit: "null",
+    quantification: "=",
+    robotoffPrediction: null
+  },
+  {
+    off_nutriment_id: "energy_kcal",
+    label: "Protein",
+    value: "",
+    unit: "null",
+    quantification: "<",
+    robotoffPrediction: null
+  },
+  {
+    off_nutriment_id: "energy_kcal",
+    label: "Shugar",
+    value: "",
+    unit: "null",
+    quantification: "<",
+    robotoffPrediction: null
+  },
+  {
+    off_nutriment_id: "energy_kcal",
+    label: "Fat",
+    value: "",
+    unit: null,
+    quantification: "<",
+    robotoffPrediction: null
+  }
+]
+
+const additionalNutrs = [
+  {
+    off_nutriment_id: "energy_kj",
+    label: "Energie (kJ)",
+    value: "",
+    unit: "null",
+    quantification: "=",
+    robotoffPrediction: null
+  },
+  {
+    off_nutriment_id: "energy_kcal",
+    label: "Protein",
+    value: "",
+    unit: "null",
+    quantification: "<",
+    robotoffPrediction: null
+  }
+]
+
+export {basicNutriments, additionalNutrs}
